refactor: code fix for modal component; delete the unUsed Modal folder and files
