+++ src/components/Modal/Modal.tsx
@@ -1,20 +1,67 @@
-import { FC } from 'react'
-import CheckRounded from '@mui/icons-material/CheckRounded'
+import { FC, ReactNode, useState } from 'react'
+import classNames from 'classnames'
+import { CgClose } from 'react-icons/cg'
+import { MdOutlineMarkEmailUnread } from 'react-icons/md'
+import { BiFile } from 'react-icons/bi'
 import CloseRoundedIcon from '@mui/icons-material/CloseRounded'
+import CheckRoundedIcon from '@mui/icons-material/CheckRounded'
 
 interface Props {
-  onClose: () => void
+  children?: ReactNode
+  variant?: IconVariant
+  onClose?: () => void
+  isError?: boolean
 }
-const Modal: FC<Props> = ({ onClose }) => (
-  <div className="fixed top-0 left-0 bottom-0 right-0 bg-background/40 flex items-center justify-center z-50">
-    <div className="rounded-[10px] w-[494px] h-[405px] pt-5 text-center bg-white shadow-md border border-background/40">
-      <div className="text-end">
-        <CloseRoundedIcon onClick={onClose} className="cursor-pointer mb-10 mr-5" />
+
+export enum IconVariant {
+  Email,
+  Validation,
+  Form,
+}
+
+const Modal: FC<Props> = ({ children = null, variant = null, onClose = undefined, isError = undefined }) => {
+  const [isClosed, setIsClosed] = useState(false)
+
+  const handleClose = (): void => {
+    setIsClosed(true)
+    if (onClose) {
+      onClose()
+    }
+  }
+
+  if (isClosed) {
+    return null
+  }
+
+  return (
+    <div className="w-screen h-screen fixed z-50 inset-0 flex items-center justify-center bg-black bg-opacity-50">
+      <div className="w-[500px] rounded bg-white p-10 min-h-[500px] flex flex-col items-center">
+        <div className="text-6xl self-end">
+          <CgClose onClick={handleClose} className="cursor-pointer" />
+        </div>
+        <div
+          className={classNames(
+            'w-24',
+            'rounded-full',
+            'flex',
+            'items-center',
+            'justify-center',
+            'my-2',
+            variant === IconVariant.Email && ['bg-gray text-white text-6xl h-24 mt-10'],
+            variant === IconVariant.Form && ['text-primary text-[100px] mt-10'],
+            variant === IconVariant.Validation && [isError ? 'bg-error' : 'bg-success', 'text-white h-24 mt-10']
+          )}
+        >
+          {variant === IconVariant.Email && <MdOutlineMarkEmailUnread />}
+          {variant === IconVariant.Validation &&
+            (isError ? <CloseRoundedIcon className="text-6xl" /> : <CheckRoundedIcon className="text-6xl" />)}
+          {variant === IconVariant.Form && <BiFile />}
+          {}
+        </div>
+        <div className="text-2xl">{children}</div>
       </div>
-      <CheckRounded className="text-white bg-[#0EA739] w-[99px] h-[99px] rounded-full" />
-      <div className="font-semi-bold text-xl mt-12">Sign Up Successfully</div>
     </div>
-  </div>
-)
+  )
+}
 
 export default Modal
